+++ src/tools/UBGraphicsTriangle.cpp
@@ -351,7 +351,7 @@ void UBGraphicsTriangle::paintGraduations(QPainter *painter)
 
     // Update the width of one "centimeter" to correspond to the width of the background grid (whether it is displayed or not)
     sPixelsPerCentimeter = UBApplication::boardController->activeScene()->backgroundGridSize();
-    double pixelsPerMillimeter = double(sPixelsPerCentimeter)/10.0;
+    double pixelsPerMillimeter = sPixelsPerCentimeter/10.0;
 
     // When a "centimeter" is too narrow, we only display every 5th number, and every 5th millimeter mark
     double numbersWidth = fontMetrics.width("00");
